segmentation/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "segmenter"
version = "0.1.0"
description = "Tibetan text segmentation pipeline with Azure integration"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.60.0",
    "openpyxl>=3.0.0",
    "azure-storage-file-share>=12.0.0",
    "botok>=0.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
segmenter = "segmenter.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["segmenter*"]
